+++ odbc/odbc_handles.h
@@ -196,6 +196,13 @@ public:
   SQLRETURN primary_keys(const std::optional<std::string>& catalog_name,
                          const std::optional<std::string>& schema_name,
                          const std::string& table_name);
+  SQLRETURN foreign_keys(
+      const std::optional<std::string>& pk_catalog_name,
+      const std::optional<std::string>& pk_schema_name,
+      const std::optional<std::string>& pk_table_name,
+      const std::optional<std::string>& fk_catalog_name,
+      const std::optional<std::string>& fk_schema_name,
+      const std::optional<std::string>& fk_table_name);
   SQLRETURN tables(const std::optional<std::string>& catalog_name,
                    const std::optional<std::string>& schema_name,
                    const std::optional<std::string>& table_name,